+++ .github/workflows/ci.yml
@@ -12,8 +12,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1'
-          - '1.6' # this is the oldest compat
+          - '1.9' # this is the oldest compat
           - 'nightly'
         os:
           - ubuntu-latest
