+++ .gitignore
@@ -1,4 +1,5 @@
 /Manifest.toml
 /docs/build/
 /docs/Manifest.toml
+/test/Manifest.toml
 _flux-theme/
