+++ test/Project.toml
@@ -1,0 +1,4 @@
+[deps]
+Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
+InfiniteArrays = "4858937d-0d70-526a-a4dd-2d5cb5dd786c"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
